maven_sys0809 ---end

+++ Aworkspace/javaweb/maven_sys/src/main/java/com/baizhi/qfl/service/CityServiceImpl.java
@@ -1,7 +1,9 @@
 package com.baizhi.qfl.service;
 
 import com.baizhi.qfl.dao.CityDao;
+import com.baizhi.qfl.dao.StudentDao;
 import com.baizhi.qfl.entity.City;
+import com.baizhi.qfl.entity.Student;
 import com.baizhi.qfl.util.DBUtil;
 import org.apache.ibatis.session.SqlSession;
 
@@ -54,7 +56,13 @@ public class CityServiceImpl implements CityService{
             session = DBUtil.openSession();
             // 调用dao的方法
             CityDao dao = session.getMapper(CityDao.class);
-            dao.delete(id);
+            StudentDao sdao=session.getMapper(StudentDao.class);
+            List<Student> list=sdao.selectByCityId(id);
+            if (list.size() != 0) {
+                 throw new RuntimeException("该城市有关联用户，不能删除");
+            }else {
+                dao.delete(id);
+            }
             session.commit();  //只有业务里涉及到 增删改
         }catch(Exception e){
             e.printStackTrace();
